Implement autosetup Implement last errors information Fix bugs

+++ Repositories/FtpAccountRepository.php
@@ -2,11 +2,20 @@
 
 namespace GameapModules\Ftp\Repositories;
 
+use Exception;
+use Gameap\Exceptions\GameapException;
+use GameapModules\Ftp\Exceptions\ExecuteCommandException;
 use GameapModules\Ftp\Models\FtpAccount;
 use GameapModules\Ftp\Services\CommandsService;
+use Cache;
 
 class FtpAccountRepository
 {
+    const EXEC_SUCCESS_CODE = 0;
+
+    const CACHE_TTL_SECONDS = 300;
+    const CACHE_LAST_ERROR_KEY = 'ftp:last_error';
+
     /**
      * @var CommandsService
      */
@@ -24,15 +33,23 @@ class FtpAccountRepository
 
     /**
      * @param array $attributes
+     * @throws ExecuteCommandException
+     * @throws GameapException
      */
     public function store(array $attributes)
     {
-        $this->commandsService->addAccount(
+        $result = $this->commandsService->addAccount(
             $attributes['ds_id'],
             $attributes['username'],
             $attributes['password'],
-            $attributes['dir']
+            $attributes['dir'],
+            $exitCode
         );
+
+        if ($exitCode !== self::EXEC_SUCCESS_CODE) {
+            Cache::put(self::CACHE_LAST_ERROR_KEY, $result, self::CACHE_TTL_SECONDS);
+            throw new ExecuteCommandException();
+        }
 
         FtpAccount::create($attributes);
     }
@@ -40,32 +57,56 @@ class FtpAccountRepository
     /**
      * @param int $id
      * @param array $attributes
+     * @throws ExecuteCommandException
+     * @throws GameapException
      */
     function update(int $id, array $attributes)
     {
         $ftpAccount = FtpAccount::findOrFail($id);
 
-        $this->commandsService->updateAccount(
+        $result = $this->commandsService->updateAccount(
             $ftpAccount->ds_id,
-            $attributes['username'],
+            $attributes['username'] ?? $ftpAccount->username,
             $attributes['password'],
-            $attributes['dir']
+            $attributes['dir'],
+            $exitCode
         );
+
+        if ($exitCode !== self::EXEC_SUCCESS_CODE) {
+            Cache::put(self::CACHE_LAST_ERROR_KEY, $result, self::CACHE_TTL_SECONDS);
+            throw new ExecuteCommandException();
+        }
 
         $ftpAccount->update($attributes);
     }
 
     /**
      * @param FtpAccount $ftpAccount
-     * @throws \Exception
+     * @throws ExecuteCommandException
+     * @throws GameapException
+     * @throws Exception
      */
     function destroy(FtpAccount $ftpAccount)
     {
-        $this->commandsService->deleteAccount(
+        $result = $this->commandsService->deleteAccount(
             $ftpAccount->ds_id,
-            $ftpAccount->username
+            $ftpAccount->username,
+            $exitCode
         );
 
+        if ($exitCode !== self::EXEC_SUCCESS_CODE) {
+            Cache::put(self::CACHE_LAST_ERROR_KEY, $result, self::CACHE_TTL_SECONDS);
+            throw new ExecuteCommandException();
+        }
+
         $ftpAccount->delete();
+    }
+
+    /**
+     * @return mixed
+     */
+    function lastError()
+    {
+        return Cache::get(self::CACHE_LAST_ERROR_KEY);
     }
 }
